Move common pagination code to mixin
